backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
email-validator
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings